Fix: 엔딩 페이지에서 Pin컴포넌트 import 시 props와 dnd 오류 수정
- handleUpdate, handleDelete 함수를 optional로 변경
- DragNDropProvider 컴포넌트를 추가하여 dnd 오류 수정
- 여행 수정 중 다른 페이지로 이동 시 modifyState를 readOnly로 변경
